@@ -14,16 +14,16 @@
 | written permission from the original author(s).        |
 +--------------------------------------------------------*/
 
-use CRM_Bic_ExtensionUtil as E;
+declare(strict_types = 1);
 
-require_once 'CRM/Bic/Parser/Parser.php';
+use CRM_Bic_ExtensionUtil as E;
 
 /**
  * Abstract class defining the basis for national bank info parsers
  */
 class CRM_Bic_Parser_AT extends CRM_Bic_Parser_Parser {
 
-  static $page_url = 'https://www.oenb.at/docroot/downloads_observ/sepa-zv-vz_gesamt.csv';
+  public static $page_url = 'https://www.oenb.at/docroot/downloads_observ/sepa-zv-vz_gesamt.csv';
 
   public function update() {
     // first, download the page
@@ -34,7 +34,7 @@ class CRM_Bic_Parser_AT extends CRM_Bic_Parser_Parser {
     }
 
     $banks   = [];
-    $headers = null;
+    $headers = NULL;
 
     // iterate CSV records
     foreach ($data as $line) {
@@ -43,16 +43,20 @@ class CRM_Bic_Parser_AT extends CRM_Bic_Parser_Parser {
       $fields = str_getcsv($line, ';');
 
       // skip some stuff: preamble
-      if (count($fields) < 20) continue;
+      if (count($fields) < 20) {
+        continue;
+      }
 
       if ($headers === NULL) {
         // first 'real' line should be header
         $headers = $fields;
 
-        if (  !in_array('Identnummer', $headers)
+        if (!in_array('Identnummer', $headers)
            || !in_array('Bankleitzahl', $headers)
            || !in_array('SWIFT-Code', $headers)) {
-          return $this->createParserOutdatedError(E::ts("Source file doesn't contain Identnummer/Bankleitzahl/SWIFT-Code"));
+          return $this->createParserOutdatedError(
+            E::ts("Source file doesn't contain Identnummer/Bankleitzahl/SWIFT-Code")
+          );
         }
       }
       else {
@@ -60,18 +64,20 @@ class CRM_Bic_Parser_AT extends CRM_Bic_Parser_Parser {
         $data_set = array_combine($headers, $fields);
 
         // we will only process banks with a BIC/SWIFT Code
-        if (   empty($data_set['Identnummer'])
+        if (empty($data_set['Identnummer'])
             || empty($data_set['Bankleitzahl'])
             || empty($data_set['SWIFT-Code'])
-        ) continue;
+        ) {
+          continue;
+        }
 
         // compile data set
-        $banks[$data_set['Bankleitzahl']] = array(
-            'value'       => $data_set['Bankleitzahl'],
-            'name'        => $data_set['SWIFT-Code'],
-            'label'       => $data_set['Bankenname'] ?? 'Unknown',
-            'description' => $this->getDescription($data_set)
-        );
+        $banks[$data_set['Bankleitzahl']] = [
+          'value'       => $data_set['Bankleitzahl'],
+          'name'        => $data_set['SWIFT-Code'],
+          'label'       => $data_set['Bankenname'] ?? 'Unknown',
+          'description' => $this->getDescription($data_set),
+        ];
       }
     }
 
@@ -88,8 +94,7 @@ class CRM_Bic_Parser_AT extends CRM_Bic_Parser_Parser {
    * @return string
    *   the description
    */
-  protected function getDescription($data_set)
-  {
+  protected function getDescription($data_set) {
     $description = '';
     if (!empty($data_set['Straße'])) {
       $description .= $data_set['Straße'] . ',';
@@ -103,12 +108,15 @@ class CRM_Bic_Parser_AT extends CRM_Bic_Parser_Parser {
     return $description;
   }
 
-  /*
- * Extracts the National Bank Identifier from an Austrian IBAN.
- */
+  /**
+   *
+   * Extracts the National Bank Identifier from an Austrian IBAN.
+   *
+   */
   public function extractNBIDfromIBAN($iban) {
-    return array(
-        substr($iban, 4, 5),
-    );
+    return [
+      substr($iban, 4, 5),
+    ];
   }
+
 }
