@@ -1,49 +1,65 @@
 <?php
 
-require_once 'bic.civix.php';
+declare(strict_types = 1);
 
-use \Symfony\Component\DependencyInjection\ContainerBuilder;
+// phpcs:disable PSR1.Files.SideEffects
+require_once 'bic.civix.php';
+// phpcs:enable
+
+use CRM_Bic_ExtensionUtil as E;
+use Symfony\Component\DependencyInjection\ContainerBuilder;
+
+function _bic_composer_autoload(): void {
+  if (file_exists(__DIR__ . '/vendor/autoload.php')) {
+    require_once __DIR__ . '/vendor/autoload.php';
+  }
+}
 
 /**
- * Implements hook_civicrm_container()
+ * Implements hook_civicrm_container().
  *
  * @link https://docs.civicrm.org/dev/en/latest/hooks/hook_civicrm_container/
  */
 function bic_civicrm_container(ContainerBuilder $container) {
-    if (class_exists('\Civi\Bic\ContainerSpecs')) {
-        $container->addCompilerPass(new \Civi\Bic\ContainerSpecs());
-    }
+  _bic_composer_autoload();
+
+  if (class_exists('\Civi\Bic\ContainerSpecs')) {
+    $container->addCompilerPass(new \Civi\Bic\ContainerSpecs());
+  }
 }
 
 /**
- * Implementation of hook_civicrm_config
+ * Implements hook_civicrm_config().
  */
 function bic_civicrm_config(&$config) {
+  _bic_composer_autoload();
   _bic_civix_civicrm_config($config);
 }
 
 /**
- * Implementation of hook_civicrm_install
+ * Implements hook_civicrm_install().
  */
 function bic_civicrm_install() {
-  return _bic_civix_civicrm_install();
+  _bic_civix_civicrm_install();
 }
 
 /**
- * Implementation of hook_civicrm_enable
+ * Implements hook_civicrm_enable().
  */
 function bic_civicrm_enable() {
-  return _bic_civix_civicrm_enable();
+  _bic_civix_civicrm_enable();
 }
 
 /**
+ * Implements hook_civicrm_alterAPIPersmissions().
+ *
  * Set permissions for runner/engine API call
  */
 function bic_civicrm_alterAPIPermissions($entity, $action, &$params, &$permissions) {
   // TODO: adjust to correct permission
-  $permissions['bic']['getfromiban'] = array('access CiviCRM');
-  $permissions['bic']['findbyiban']  = array('access AJAX API');
-  $permissions['bic']['get']         = array('access CiviCRM');
+  $permissions['bic']['getfromiban'] = ['access CiviCRM'];
+  $permissions['bic']['findbyiban']  = ['access AJAX API'];
+  $permissions['bic']['get']         = ['access CiviCRM'];
 }
 
 /**
@@ -53,15 +69,15 @@ function bic_civicrm_alterAPIPermissions($entity, $action, &$params, &$permissio
  *
  */
 function bic_civicrm_navigationMenu(&$menu) {
-  _bic_civix_insert_navigation_menu($menu, 'Search', array(
-    'label' => ts('Find Banks', array('domain' => 'org.project60.bic')),
+  _bic_civix_insert_navigation_menu($menu, 'Search', [
+    'label' => E::ts('Find Banks', ['domain' => 'org.project60.bic']),
     'name' => 'BankLists',
     'url' => 'civicrm/bicList',
     'permission' => 'access CiviContribute',
     'operator' => NULL,
     'separator' => 2,
     'active' => 1,
-  ));
+  ]);
 
   _bic_civix_navigationMenu($menu);
 }
