@@ -14,35 +14,38 @@
 | written permission from the original author(s).        |
 +--------------------------------------------------------*/
 
+declare(strict_types = 1);
+
 use CRM_Bic_ExtensionUtil as E;
 
 /**
  * Collection of upgrade steps.
  */
-class CRM_Bic_Upgrader extends CRM_Extension_Upgrader_Base
-{
+class CRM_Bic_Upgrader extends CRM_Extension_Upgrader_Base {
 
-    /**
-     * Make sure the option group is there
-     */
-    public function enable()
-    {
-        // check if option group is there, and create if it isn't
-        try {
-            $option_group = civicrm_api3('OptionGroup', 'getsingle', array('name' => 'bank_list'));
-        } catch (Exception $e) {
-            // group's not there yet, create:
-            try {
-                $option_group = civicrm_api3('OptionGroup', 'create', array(
-                    'name'        => 'bank_list',
-                    'title'       => ts('List of banks'),
-                    'is_reserved' => 0,
-                    'is_active'   => 1,
-                ));
-            } catch (Exception $create_ex) {
-                // TODO: more info?
-                Civi::log()->warning("Couldn't create 'bank_list' OptionGroup.");
-            }
-        }
+  /**
+   * Make sure the option group is there
+   */
+  public function enable() {
+    // check if option group is there, and create if it isn't
+    try {
+      $option_group = civicrm_api3('OptionGroup', 'getsingle', ['name' => 'bank_list']);
     }
+    catch (Exception $e) {
+      // group's not there yet, create:
+      try {
+        $option_group = civicrm_api3('OptionGroup', 'create', [
+          'name'        => 'bank_list',
+          'title'       => ts('List of banks'),
+          'is_reserved' => 0,
+          'is_active'   => 1,
+        ]);
+      }
+      catch (Exception $create_ex) {
+        // TODO: more info?
+        Civi::log()->warning("Couldn't create 'bank_list' OptionGroup.");
+      }
+    }
+  }
+
 }
